+++ lib/kubecontrol/resources/secret.rb
@@ -1,3 +1,6 @@
+require 'json'
+require "base64"
+
 module Kubecontrol
   module Resources
     class Secret
@@ -12,6 +15,18 @@ module Kubecontrol
         @age = age
         @namespace = namespace
         @client = client
+      end
+
+      def data_values
+        @data_values ||= begin
+                           std_out, _std_err, exit_code = @client.kubectl_command("get secret #{@name} -o json")
+                           if exit_code.zero?
+                             json_secret = JSON.parse(std_out)
+                             json_secret['data'].reduce({}) {|h, (k,v)|  h[k] = Base64.decode64(v); h }
+                           else
+                             {}
+                           end
+                         end
       end
     end
   end